started implementing basic can protocol

+++ include/config.h
@@ -39,7 +39,7 @@ struct Config
     bool CAN_ENABLED;
     uint8_t PIN_CAN_RX;
     uint8_t PIN_CAN_TX;
-    uint16_t CAN_SPEED_KBPS;
+    uint8_t CAN_SPEED_KBPS;
     uint8_t PIN_ENABLE_CONFIG;
     size_t PARAMETERS_LENGTH;
     ParameterDef** PARAMETERS;
